Update AddLand, ProfileSettings, ProfileCRUD and AuthUser. Update database tables
Update register and update to include phone number and removed the role from both functions and the database schema. Update verifications email to be redirecting to the login page and for reset password updated the email to be taking directly to the front end route with a token. add land now toggless weather api without the need to do it seperately. add humidity to update land. To do in the future : whenever we update land we re generate business plan (until we fix flask deployment)